Improve user experience
User experience (UX) refers to a person's total experience using a
particular product, system or service.

The first requirement for a great user experience is to meet the exact
needs for the usage of a product or a service, without fuss or bother.
To achieve this, users must have all necessary information for their
task without being overloaded, not having unnecessary actions,
calculation or having to recall. Interaction techniques must be
familiar and behave along expectation, errors must be prevented. If
users make errors, there must be easy way to recover them. When usage
is frequent, product or service must provide shortcut. Users need to
have feedbacks, feel in control and know what is going on at all time.

User experience includes the practical, experiential, affective,
meaningful and valuable aspects of human–computer interaction and
product ownership. Additionally, it includes a person’s perceptions of
system aspects such as utility, ease of use and efficiency. User
experience may be considered subjective in nature to the degree that it
is about individual perception and thought with respect to the system.
User experience is dynamic as it is constantly modified over time due
to changing usage circumstances and changes to individual systems as
well as the wider usage context in which they can be found.