Alternate single- and double- quotes in prefs array
Previously, the pref strings were sometimes single-quoted and sometimes
double-quoted, with no clear pattern. This is confusing to the
programmer, me, and may lead said programmer, me, to erroneously think
that there is a semantic difference between single- and double- quoted
strings, when no such difference exists in reality.

This issue was resolved by alternating between single- and double-
quotes, so that a clear pattern was established. This preserves the
original combination of single- and double- quotes in the prefs array,
while making the decision of whether to single-quote or double-quote a
new addition to the array clear to the programmer – assuming that the
programmer is appending an entry to the array, said programmer (you, me,
or other), must only look at the previous string, and use the alternate
quoting style on the new string.

One caveat to this approach is that insertion of a single new item into
the array (not at the beginning or end) while preserving the quotation
pattern is an O(n) operation on the part of the programmer. For maximum
productivity and since the ordering of items in the array does not
currently matter, developers are encouraged to insert new items at the
beginning or end of the array, since either type of insertion is an O(1)
operation.